requirements.txt
pyTelegramBotAPI
datetime
re
mysql-connector-python